@@ -36,7 +36,7 @@
     "babel-preset-es2016": "^6.22.0",
     "babel-preset-react": "^6.22.0",
     "conventional-changelog-cli": "~1.2.0",
-    "css-loader": "^0.26.1",
+    "css-loader": "^0.27.3",
     "dom4": "^1.8.3",
     "enzyme": "^2.7.1",
     "esdoc": "^0.5.2",
@@ -45,9 +45,9 @@
     "eslint-config-airbnb": "^14.0.0",
     "eslint-loader": "^1.6.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.9.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.1.0",
     "jasmine-core": "^2.5.2",
     "karma": "^1.4.0",
     "karma-babel-preprocessor": "^6.0.1",
@@ -59,28 +59,30 @@
     "karma-webpack": "^2.0.1",
     "node-sass": "^4.3.0",
     "postcss-loader": "^1.2.2",
-    "react": "^15.4.2",
+    "react": "^15.5.4",
     "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.5.4",
     "sass-lint": "^1.10.2",
-    "sass-loader": "^4.1.1",
+    "sass-loader": "^6.0.3",
     "sasslint-webpack-plugin": "^1.0.4",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2"
+    "style-loader": "^0.14.0",
+    "uglify-js": "^2.8.16",
+    "uglifyjs-webpack-plugin": "^0.3.1",
+    "webpack": "^2.2.1",
+    "webpack-dev-server": "^2.4.2"
   },
   "peerDependencies": {
     "react": "^15.0.0",
     "react-dom": "^15.0.0"
   },
   "scripts": {
-    "prebuild": "npm run test && rm -rf lib",
+    "prebuild": "npm test && rm -rf lib",
     "build": "npm run build:lib && npm run build:scss && npm run build:bundle",
     "build:bundle": "NODE_ENV=production webpack",
     "build:lib": "NODE_ENV=production babel src/js --out-dir lib/js --source-maps",
     "build:scss": "NODE_ENV=production node-sass src/scss --recursive --source-map true --output lib/css",
     "changelog": "conventional-changelog --preset angular --infile CHANGELOG.md --same-file",
-    "dev": "webpack-dev-server --inline --config webpack-example.config.js --content-base example",
+    "dev": "webpack-dev-server --inline --config webpack-example.config.babel.js --content-base example",
     "docs": "esdoc",
     "lint": "npm run lint:js && npm run lint:scss",
     "lint:js": "eslint src/js/ --ext .js --ext .jsx",
@@ -90,6 +92,7 @@
     "test": "karma start --single-run"
   },
   "dependencies": {
-    "autobind-decorator": "^1.3.4"
+    "autobind-decorator": "^1.3.4",
+    "prop-types": "^15.5.8"
   }
 }
